Use BenchmarkDotNet
