+++ performance-joins/performance-joins/Building.cs
@@ -2,7 +2,7 @@
 
 namespace performance_joins
 {
-    class Building
+    public class Building
     {
         public string Id { get; set; }
         public List<string> Workers { get; set; } = new List<string>();
