+++ performance-joins/performance-joins/Worker.cs
@@ -1,6 +1,6 @@
 ﻿namespace performance_joins
 {
-    class Worker
+    public class Worker
     {
         public string Id { get; set; }
         public bool IsEmployed { get; set; }
